requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.2
gunicorn==21.2.0